+++ src/Metronome/controls/ControlPanelItem.tsx
